Move linkage packets
