+++ src/communication/linkage/index.ts
@@ -1,4 +1,2 @@
 export * from './comm-link';
 export * from './comm-linker';
-export * from './comm-messaging-request';
-export * from './message-comm-link';
