+++ src/impl/formation/index.ts
@@ -1,1 +1,2 @@
+export * from './comm-messaging-request';
 export * from './formation.comm-linker';
