+++ src/impl/hub/index.ts
@@ -1,1 +1,2 @@
 export * from './hub.comm-linker';
+export * from './message-comm-link';
